chore: decompose monolithic files and implement bit-bucket optimization
- Decomposed StorageEngine, StorageOps, and Orchestrator handlers into smaller, focused modules to enforce the 400-line maintainability limit (Issue #184).
- Implemented Bit-Bucket virtual file handle (BEP 47) to avoid physical disk usage for BitTorrent padding files (Issue #183).
- Synchronized ARCHITECTURE.md with the actual source tree paths and memory model (Issue #186).
- Updated ADR 0019 to reflect the removal of the shallow BufferPool abstraction.
- Verified all changes with the Green Loop (fmt, clippy, test, bench).
