+++ aura-core/src/storage/mod.rs
@@ -1,7 +1,10 @@
+pub mod completion;
 pub mod engine;
 pub mod ops;
+pub mod registry;
 pub mod scheduler;
 pub mod sys;
+pub mod utils;
 pub use engine::*;
 
 #[cfg(test)]
