+++ aura-core/src/storage/ops.rs
@@ -1,31 +1,47 @@
+use super::utils::get_part_path;
 use super::StorageEngine;
 use crate::worker::Segment;
 use crate::Error;
 use crate::{Result, TaskId};
 use bytes::{Bytes, BytesMut};
-use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use tokio::fs::{self, File, OpenOptions};
-use tracing::{debug, info};
+use tracing::debug;
 
 impl StorageEngine {
     pub(crate) async fn handle_read(&mut self, id: TaskId, segment: Segment) -> Result<Bytes> {
-        let buffer_size = self
-            .config
-            .as_ref()
-            .map(|cfg: &Arc<arc_swap::ArcSwap<crate::Config>>| {
-                cfg.load().storage.read_ahead_kb as usize * 1024
-            })
-            .unwrap_or(128 * 1024); // Default to 128 KB seeding read buffer capacity
-
-        let file = self.get_or_open_part_file(id).await?;
+        let file: &mut File = self.get_or_open_part_file(id).await?;
         use tokio::io::AsyncReadExt;
         use tokio::io::AsyncSeekExt;
+
+        let file_len = file.metadata().await?.len();
+        let read_len = std::cmp::min(segment.length, file_len.saturating_sub(segment.offset));
+
         file.seek(std::io::SeekFrom::Start(segment.offset)).await?;
 
-        let mut reader = tokio::io::BufReader::with_capacity(buffer_size, file);
         let mut buf = vec![0u8; segment.length as usize];
-        reader.read_exact(&mut buf).await?;
+        if read_len > 0 {
+            file.read_exact(&mut buf[..read_len as usize]).await?;
+        }
+
+        // Explicitly zero out padding ranges in the buffer
+        let padding_ranges = self
+            .task_padding_ranges
+            .get(&id)
+            .cloned()
+            .unwrap_or_default();
+        let sub_pads = super::engine::get_padding_subranges_impl(
+            &padding_ranges,
+            segment.offset,
+            segment.length,
+        );
+
+        for pad in sub_pads {
+            let start_in_buf = (pad.start - segment.offset) as usize;
+            let end_in_buf = (pad.end - segment.offset) as usize;
+            buf[start_in_buf..end_in_buf].fill(0);
+        }
+
         Ok(Bytes::from(buf))
     }
 
@@ -40,12 +56,44 @@ impl StorageEngine {
         if segment.offset == next_offset {
             let len = data.len() as u64;
 
-            // Push to dirty buffer instead of immediate write
-            if let Some(dirty) = self.dirty_buffers.get_mut(&id) {
-                dirty.push((segment.offset, data));
-            }
-            if let Some(size) = self.dirty_sizes.get_mut(&id) {
-                *size += len as usize;
+            let padding_ranges = self
+                .task_padding_ranges
+                .get(&id)
+                .cloned()
+                .unwrap_or_default();
+
+            // Filter padding before pushing to dirty buffer
+            let subranges =
+                super::engine::get_non_padding_subranges_impl(&padding_ranges, segment.offset, len);
+            for sub in subranges {
+                let sub_offset = sub.start;
+                let sub_len = sub.length();
+
+                // Check for discontinuity in dirty buffer (caused by padding skip)
+                let needs_flush = if let Some(dirty) = self.dirty_buffers.get(&id) {
+                    if let Some(last) = dirty.last() {
+                        sub_offset != (last.0 + last.1.len() as u64)
+                    } else {
+                        false
+                    }
+                } else {
+                    false
+                };
+
+                if needs_flush {
+                    self.flush_dirty_buffer(id).await?;
+                }
+
+                let start_in_data = (sub_offset - segment.offset) as usize;
+                let end_in_data = start_in_data + sub_len as usize;
+                let sub_data = BytesMut::from(&data[start_in_data..end_in_data]);
+
+                if let Some(dirty) = self.dirty_buffers.get_mut(&id) {
+                    dirty.push((sub_offset, sub_data));
+                }
+                if let Some(size) = self.dirty_sizes.get_mut(&id) {
+                    *size += sub_len as usize;
+                }
             }
 
             let mut current_offset = next_offset + len;
@@ -54,13 +102,41 @@ impl StorageEngine {
             if let Some(pending) = self.pending_writes.get_mut(&id) {
                 while let Some(p_data) = pending.remove(&current_offset) {
                     let p_len = p_data.len() as u64;
-                    to_flush.push((current_offset, p_data));
+
+                    // Filter padding for pending writes as well
+                    let p_subranges = super::engine::get_non_padding_subranges_impl(
+                        &padding_ranges,
+                        current_offset,
+                        p_len,
+                    );
+                    for sub in p_subranges {
+                        let start_in_p = (sub.start - current_offset) as usize;
+                        let end_in_p = start_in_p + sub.length() as usize;
+                        to_flush.push((sub.start, BytesMut::from(&p_data[start_in_p..end_in_p])));
+                    }
+
                     current_offset += p_len;
                 }
             }
 
             for (offset, p_data) in to_flush {
                 let p_len = p_data.len();
+
+                // Check for discontinuity again
+                let needs_flush = if let Some(dirty) = self.dirty_buffers.get(&id) {
+                    if let Some(last) = dirty.last() {
+                        offset != (last.0 + last.1.len() as u64)
+                    } else {
+                        false
+                    }
+                } else {
+                    false
+                };
+
+                if needs_flush {
+                    self.flush_dirty_buffer(id).await?;
+                }
+
                 if let Some(dirty) = self.dirty_buffers.get_mut(&id) {
                     dirty.push((offset, p_data));
                 }
@@ -146,7 +222,7 @@ impl StorageEngine {
     }
 
     pub(crate) async fn execute_io_task(&mut self, task: super::scheduler::IoTask) -> Result<()> {
-        let file = self.get_or_open_part_file(task.task_id).await?;
+        let file: &mut File = self.get_or_open_part_file(task.task_id).await?;
         use tokio::io::AsyncSeekExt;
         use tokio::io::AsyncWriteExt;
 
@@ -159,125 +235,6 @@ impl StorageEngine {
 
         crate::storage::sys::apply_fadvise_dontneed(file, task.offset, total_len);
 
-        Ok(())
-    }
-
-    pub(crate) async fn handle_complete(&mut self, id: TaskId) -> Result<()> {
-        let base_path = self.task_paths.get(&id).ok_or(Error::TaskNotFound(id))?;
-        let part_path = get_part_path(base_path)?;
-        let hardened_base = crate::storage::sys::harden_path(base_path);
-
-        self.flush_all_pending(id).await?;
-
-        let tasks = self.scheduler.extract_all_for_task(id);
-        for task in tasks {
-            self.execute_io_task(task).await?;
-        }
-
-        // Close the write handle before verification to ensure all data is committed
-        // and to allow clean read-only access.
-        self.handles.pop(&id);
-
-        // fsync the .part file to ensure all data is on disk before exposing it under the final name
-        let file = fs::OpenOptions::new().read(true).open(&part_path).await?;
-        file.sync_all().await?;
-
-        info!(%id, from = ?part_path, to = ?hardened_base, "Performing atomic completion rename");
-        fs::rename(&part_path, &hardened_base).await?;
-
-        // Sync parent directory to ensure metadata rename is durable on Unix
-        sync_parent_dir(&hardened_base).await;
-
-        let _ = self
-            .completion_tx
-            .send(crate::storage::StorageEvent::Completed(id))
-            .await;
-
-        Ok(())
-    }
-
-    pub(crate) async fn verify_checksum(&mut self, id: TaskId) -> Result<()> {
-        let checksum = match self.task_checksums.get(&id) {
-            Some(c) => c.clone(),
-            None => return Ok(()),
-        };
-
-        info!(%id, ?checksum, "Verifying file integrity");
-
-        let base_path = self.task_paths.get(&id).ok_or(Error::TaskNotFound(id))?;
-        let part_path = get_part_path(base_path)?;
-
-        let file = File::open(&part_path).await?;
-        let mut reader = tokio::io::BufReader::new(file);
-
-        use md5::Digest;
-        use tokio::io::AsyncReadExt;
-
-        let actual = match checksum {
-            crate::Checksum::Md5(ref expected) => {
-                let mut hasher = md5::Md5::default();
-                let mut buffer = [0u8; 65536];
-                loop {
-                    let n = reader.read(&mut buffer).await?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buffer[..n]);
-                }
-                let hash = hex::encode(hasher.finalize());
-                (expected.clone(), hash)
-            }
-            crate::Checksum::Sha1(ref expected) => {
-                let mut hasher = sha1::Sha1::default();
-                let mut buffer = [0u8; 65536];
-                loop {
-                    let n = reader.read(&mut buffer).await?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buffer[..n]);
-                }
-                let hash = hex::encode(hasher.finalize());
-                (expected.clone(), hash)
-            }
-            crate::Checksum::Sha256(ref expected) => {
-                let mut hasher = sha2::Sha256::default();
-                let mut buffer = [0u8; 65536];
-                loop {
-                    let n = reader.read(&mut buffer).await?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buffer[..n]);
-                }
-                let hash = hex::encode(hasher.finalize());
-                (expected.clone(), hash)
-            }
-            crate::Checksum::Sha512(ref expected) => {
-                let mut hasher = sha2::Sha512::default();
-                let mut buffer = [0u8; 65536];
-                loop {
-                    let n = reader.read(&mut buffer).await?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buffer[..n]);
-                }
-                let hash = hex::encode(hasher.finalize());
-                (expected.clone(), hash)
-            }
-        };
-
-        let (expected, actual_hash) = actual;
-
-        if expected.to_lowercase() != actual_hash.to_lowercase() {
-            return Err(Error::Storage(format!(
-                "Checksum mismatch: expected {}, got {}",
-                expected, actual_hash
-            )));
-        }
-
-        info!(%id, "Integrity verification successful");
         Ok(())
     }
 
@@ -310,27 +267,4 @@ impl StorageEngine {
             .get_mut(&id)
             .expect("File handle must exist after open/insert"))
     }
-}
-
-async fn sync_parent_dir(path: &Path) {
-    if let Some(parent) = path.parent() {
-        let parent_clone = parent.to_path_buf();
-        let _ = tokio::task::spawn_blocking(move || {
-            if let Ok(dir) = std::fs::File::open(&parent_clone) {
-                let _ = dir.sync_all();
-            }
-        })
-        .await;
-    }
-}
-
-pub fn get_part_path(base_path: &Path) -> Result<PathBuf> {
-    let mut part_path = crate::storage::sys::harden_path(base_path);
-    let mut filename = part_path
-        .file_name()
-        .ok_or_else(|| Error::Task(TaskId(0), "Invalid filename".to_string()))? // Placeholder ID as we don't have it here
-        .to_os_string();
-    filename.push(".part");
-    part_path.set_file_name(filename);
-    Ok(part_path)
 }
